ppe_monitor/requirements.txt
numpy==2.0.2
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
streamlit==1.40.2
torch==2.5.1
transformers==4.46.2
ultralytics==8.3.32
facenet-pytorch==2.5.3
inference-gpu==0.32.0
supervision==0.25.1